End-To-End Projects/Medical Chatbot/store_index.py
from src.helper import load_pdf_file, text_split, init_embeddings
from pinecone.grpc import PineconeGRPC as Pinecone
from pinecone import ServerlessSpec
from langchain_pinecone import PineconeVectorStore
from dotenv import load_dotenv
import os

load_dotenv()

os.environ["PINECONE_API_KEY"] = os.getenv("PINECONE_API_KEY")

extracted_data = load_pdf_file(data="Data/")
text_chunks = text_split(extracted_data=extracted_data)
embeddings = init_embeddings()

pc = Pinecone()

index_name = "medical-chatbot"

pc.create_index(
    name=index_name,
    dimension=768,
    metric="cosine",
    spec=ServerlessSpec(
        cloud="aws",
        region="us-east-1",
    )
)

docsearch = PineconeVectorStore.from_documents(
    documents=text_chunks,
    index_name=index_name,
    embedding=embeddings,
)

